Fix sidebar toggle button when using VoiceOver on iOS
When using iOS with VoiceOver enabled, double-tapping the sidebar toggle
button caused the sidebar to open and then immediately close. The
immediate close was caused by the `onElementTouchStart` handler in
`guest.coffee` that is intended to hide the sidebar when the user taps
somewhere in the document outside the sidebar. This handler did not
check however whether the event originated from annotator UI or not.

The fix here is to ignore any events which occur inside the sidebar
frame's container.

Fixes #1765